updated fyEventController
